Rename packages sample-compose
