+++ samples/compose-kotlin/src/main/java/nl/dionsegijn/sample/compose/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package nl.dionsegijn.demo_compose.ui.theme
+package nl.dionsegijn.sample.compose.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
